# shakira-social-channel-openapi-addendum.yaml
openapi: 3.1.0
info:
  title: Shakira Beauty Social Channel API Addendum
  version: 1.0.0
  summary: WhatsApp + Instagram endpoints layered on top of the core booking API.
servers:
  - url: https://api.shakirabeauty.example.com/v1
security:
  - bearerAuth: []
...
paths:
...
  /webhooks/whatsapp:
    post:
      tags: [Webhooks]
      security: []
      summary: Receive WhatsApp webhook events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '204':
          description: Accepted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...